devops/scripts/azd/helpers/requirements-cosmos.txt
pymongo>=4.6,<5
azure-identity>=1.15.0
dotenv